+++ eminus/extras/dftd3.py
@@ -1,0 +1,69 @@
+#!/usr/bin/env python3
+"""DFT-D3 dispersion correction interface."""
+from dftd3.interface import (
+    DispersionModel,
+    ModifiedRationalDampingParam,
+    ModifiedZeroDampingParam,
+    OptimizedPowerDampingParam,
+    RationalDampingParam,
+    ZeroDampingParam,
+)
+import numpy as np
+
+from ..data import SYMBOL2NUMBER
+from ..logger import log
+
+
+def get_Edisp(scf, version='d3bj', atm=True, xc=None):
+    """Calculate the DFT-D3 dispersion correction energy.
+
+    Reference: J. Chem. Phys. 132, 154104.
+
+    Args:
+        scf: SCF object.
+
+    Keyword Args:
+        version (str): Damping function, can be one of:
+            d3bj', 'd3bjm', 'd3zero', 'd3zerom', or 'd3op'.
+        atm (bool): Weather to use three-body dispersion energies.
+        xc (str | None): Overwrite the functional keyword if the automatic detection does not work.
+
+    Returns:
+        float: Dispersion correction energy.
+    """
+    # Dictionary of implemented dispersion models
+    dispersion_version = {
+        'd3bj': RationalDampingParam,
+        'd3bjm': ModifiedRationalDampingParam,
+        'd3zero': ZeroDampingParam,
+        'd3zerom': ModifiedZeroDampingParam,
+        'd3op': OptimizedPowerDampingParam
+    }
+
+    # Set up input parameters
+    # The dispersion correction is only geometry dependent which makes handling the input easy
+    atoms = scf.atoms
+    positions = atoms.X
+    numbers = np.array([SYMBOL2NUMBER[ia] for ia in atoms.atom])
+    # Try to determine the method keyword
+    if xc is None:
+        if scf.xc_type == 'lda':
+            log.warning('Dispersion correction for LDA functionals is not supported.')
+            return 0
+        elif 'pbe_sol' in ''.join(scf.xc):
+            method = 'pbesol'
+        elif 'pbe' in ''.join(scf.xc):
+            method = 'pbe'
+        else:
+            method = 'pbe'
+            log.warning('Functional for the dispersion correction could not be detected, continue'
+                         ' with pbe. You may need to overwrite the functional keyword manually.')
+    else:
+        method = xc
+
+    # Set up the dispersion model object with periodic boundary conditions
+    model = DispersionModel(numbers, positions, lattice=atoms.R, periodic=np.array([1, 1, 1]))
+    # Calculate the dispersion energy, neglecting the gradient
+    res = model.get_dispersion(dispersion_version[version](method=method, atm=atm), grad=False)
+    scf.energies.Edisp = res['energy']
+    return res['energy']